requirements.txt
flask
sqlalchemy==1.3.23
flask_login
flask_migrate
flask_wtf
flask-sqlalchemy>=2.5
email_validator
python-decouple
gunicorn
fiona
pyproj
rtree
shapely
geopandas
plotly
matplotlib
tqdm
qrcode
pyotp
networkx
sphinx==3.5.4
sphinx_rtd_theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4